refactor:  customProvider 改名为 providerRegistry
