+++ packages/fluent-editor/src/modules/collaborative-editing/provider/index.ts
@@ -1,3 +1,3 @@
-export * from './customProvider'
+export * from './providerRegistry'
 export * from './webrtc'
 export * from './websocket'
